requirements.txt
flask==3.1.1
flask-cors==6.0.0
requests==2.32.3
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.1.0
chromadb==0.4.24
transformers==4.36.2
torch>=2.2.0
sentence-transformers==2.2.2
